Incorporate types of title formats
Previously this PR has been focussed on mapping the data structures to
make the title format editor workable. This patch incorporates the
various types of titles we could encounter, such as for a front page,
archive page, post, etc...

Now, when clicking on the segmented control, the editor should load the
appropriate and corresponding title format. Similarly, it should load
the list of suggestions with the appropriate available tokens.